extract getUsername() to common.ts
