+++ src/common.ts
@@ -1,6 +1,12 @@
 import { readFileSync } from "fs";
 import * as path from "path";
-import { AutocompleteInteraction, CommandInteraction, TextChannel } from "discord.js";
+import {
+  AutocompleteInteraction,
+  CommandInteraction,
+  CommandInteractionOption,
+  GuildMember,
+  TextChannel,
+} from "discord.js";
 import { RequestError, getLanguages, Language } from "./codewars";
 import fuzzysearch from "fuzzysearch";
 
@@ -17,6 +23,23 @@ export const getTexts = (commandName: string, values: string[]): Map<string, str
     process.exit(1);
   }
   return texts;
+};
+
+/**
+ * Attemps to parse a username from the 'username' option. If no value is present the display name of the user sending the interaction is taken.
+ * @param interaction the CommandInteraction to get the values from
+ * @returns username
+ * @throws RequestError if the username could not be fetched
+ */
+export const getUsername = (interaction: CommandInteraction): string => {
+  let username = interaction.options.getString("username");
+  if (!username) {
+    const member = interaction.member;
+    const displayName = member instanceof GuildMember ? member.displayName : member?.nick;
+    if (!displayName) throw new RequestError("Failed to fetch the name of the current user");
+    username = displayName;
+  }
+  return username;
 };
 
 /**
